fix bug about 'no enough free place'
allow to go to Gamelauncher screen without adding player if some player are yet
registered locally :
a lock is possible if the max player is reached and a back is done
toward allplayers screen ... the Gamelauncher cannot be reached, then
the game cannot be launched.
The solution must be improved later.